Fixing firestore deploy on fresh projects
